% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Add together two numbers.}
\usage{
hello()
}
\arguments{
\item{x}{A number.}

\item{y}{A number.}
}
\value{
The sum of \code{x} and \code{y}. test
}
\description{
Add together two numbers.
}
\examples{
add(1, 1)
add(10, 1)
}

