% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmpa.R
\name{mmpa}
\alias{mmpa}
\title{Mini Pooling algorithm with covariates}
\usage{
mmpa(v, s, K = 5, vf_cut = 1000, lod = 0, method = "permutation",
  perm_num = 100)
}
\arguments{
\item{v}{Vector of viral load.}

\item{s}{Vector of risk score with the same length of viral load.}

\item{K}{Pool size, default is 5.}

\item{vf_cut}{Cutoff for individual viral failure, default is 1000.}

\item{lod}{Vector of true VL of those undetectable, default is 0.}

\item{method}{The method used for the average test number calculation, default is "permutation".}

\item{perm_num}{The number of permutation to be used for the calculation, default is 100.}
}
\value{
The average number of assays needed per pool and per subject.
}
\description{
This function allows you to compute the average number of assays needed per pool using mini pooling algorithm with covariates.
}
\examples{
d = Simdata
V = d$VL # Viral Load
S = d$S # Risk Score
mmpa(V, S, K = 3, perm_num = 3)
foo; table(foo)
}
\keyword{Pooling.}

