% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling-basic-funs.R
\name{mmpa0}
\alias{mmpa0}
\title{Number of Assays Need for Marker-Assisted Mini-Pooling with Algorithm}
\usage{
mmpa0(v, s, K = 5, vf_cut = 500, lod = 0)
}
\arguments{
\item{v}{Vector of numerical assay results.}

\item{s}{Vector of risk score with the same length of viral load.}

\item{K}{Pool size, default is \code{K=5}.}

\item{vf_cut}{Cutoff for defining disease positive, default is \code{vf_cut = 1000}.}

\item{lod}{Vector of lower limited of detection, default is \code{lod = 0}.}
}
\value{
The average number of assays needed per pool and per subject.
}
\description{
This function uses Monte Carlo to compute the average number of assays
needed per pool using mMPA.
}
\details{
alksdf detials
}
\examples{
d = Simdata
V = d$VL # Viral Load
S = d$S # Risk Score
mmpa(V, S, K = 3, perm_num = 3)
foo; table(foo)
}
\keyword{Pooling.}
