% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mMPA.R
\docType{package}
\name{mMPA}
\alias{mMPA}
\alias{mMPA-package}
\title{mMPA: A package for implementing marker-assisted mini-pooling
with algorithm}
\description{
The \code{mMPA} package provides three categories of important functions:
foo, bar and baz.
}
\details{
Add some details hers (take from paper abstract)
}
\section{functions}{

The foo functions ...
}

