% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling-basic-funs.R
\name{mmpa0}
\alias{mmpa0}
\title{Number of Assays Needed using Marker-Assisted Mini-Pooling with Algorithm}
\usage{
mmpa0(v, s, K = 5, vf_cut = 500, lod = 0)
}
\arguments{
\item{v}{Vector of numerical assay results.}

\item{s}{Vector of risk score with the same length of viral load.}

\item{K}{Pool size, default is \code{K=5}.}

\item{vf_cut}{Cutoff for defining disease positive, default is \code{vf_cut = 1000}.}

\item{lod}{Vector of lower limited of detection, default is \code{lod = 0}.}
}
\value{
A vectorof length \code{N\%/\%K} for the numbers of assays needed for all pools
that are formed.
}
\description{
Function \code{mmpa0} calculates the number of assays when using mMPA.
The pools are formed following the order of individual samples in the data.
}
\details{
For a pool size of \code{K}, the first \code{K} samples are combined to
form a pool, the next \code{K} samples are combined to form the second
pool, and so on. If the number of samples for the last pool is less than
\code{K}, these samples are not used to form a pool (i.e. not included)
in the calculation. Therefore for \code{N} samples, a total of
\code{N\%/\%K} pools are formed. The function calculates the number of
assays needed for each pool.
}
\examples{
d = Simdata
V = d$VL # Viral Load
S = d$S # Risk Score
mmpa(V, S, K = 3, perm_num = 3)
foo; table(foo)

}
\keyword{Pooling.}
