% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooling-basic-funs.R
\name{pooling_mc}
\alias{pooling_mc}
\title{Monte Carlo Simulation for Estimating the Number of Assays Needed
when Using mMPA}
\usage{
pooling_mc(v, s, K = 5, vf_cut = 1000, lod = 0, perm_num = 100,
  msg = F)
}
\arguments{
\item{perm_num}{The number of permutation to be used for the calculation;
default is \code{100}.}

\item{msg}{Message generated during calculation; default is \code{FALSE}.}
}
\value{
The outcome is a matrix of dimension \code{num_pool} by \code{perm_num}.
The row number is the number of pools (\code{num_pool}) for each permutation,
which is
determined by the sample size \code{N} and pool size \code{K}; \code{num_pool
= N\%/\%K}. The column number is the number of
permutations (\code{num_pool}).
}
\description{
This function uses Monte Carlo to simulate different orders in
which the samples would be collected to form pools. Unlike the
function \code{mmpa0} which calculates the number of assays
needed for pools that are formed following the exact order
of the samples that are listed in the data, the function
\code{mmp} permutes the data many (\code{perm_num}) times
and thus can be used to estimate the average number of
assays required (ATR) per individual without depending on a
specific configuration of forming pools. To obtain a reliable
estimate of ATR, a large number of \code{perm_num} should be used.
}
\details{
See \link{mmpa0} for detail descriptions
}
\examples{
d = Simdata
V = d$VL # Viral Load
S = d$S # Risk Score
mmpa(V, S, K = 3, perm_num = 3)
foo; table(foo)
}
\seealso{
\link{mmpa0}
}
\keyword{Pooling.}
